sandbox_api/requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
websockets>=10.1
python-multipart>=0.0.5
python-dotenv>=0.19.0
aiofiles>=0.7.0
python-socketio[asyncio]>=5.5.0